Create signers module
