+++ tests/security/test_pausable.py
@@ -2,8 +2,9 @@ import pytest
 from starkware.starknet.testing.starknet import Starknet
 from utils import (
     TRUE, FALSE, assert_revert, assert_event_emitted, 
-    get_contract_def, cached_contract, TestSigner
+    get_contract_def, cached_contract
 )
+from signers import TestSigner
 
 signer = TestSigner(12345678987654321)
