+++ tests/token/erc20/test_ERC20_Mintable.py
@@ -1,10 +1,11 @@
 import pytest
 from starkware.starknet.testing.starknet import Starknet
 from utils import (
-    TestSigner, to_uint, add_uint, sub_uint, str_to_felt, 
+    to_uint, add_uint, sub_uint, str_to_felt, 
     MAX_UINT256, ZERO_ADDRESS, INVALID_UINT256, get_contract_def, 
     cached_contract, assert_revert, assert_event_emitted
 )
+from signers import TestSigner
 
 signer = TestSigner(123456789987654321)
