+++ tests/token/erc20/test_ERC20_Pausable.py
@@ -1,9 +1,10 @@
 import pytest
 from starkware.starknet.testing.starknet import Starknet
 from utils import (
-    TestSigner, TRUE, FALSE, to_uint, str_to_felt, assert_revert, 
+    TRUE, FALSE, to_uint, str_to_felt, assert_revert, 
     get_contract_def, cached_contract
 )
+from signers import TestSigner
 
 signer = TestSigner(123456789987654321)
