+++ tests/token/erc721/test_ERC721_Mintable_Pausable.py
@@ -1,9 +1,10 @@
 import pytest
 from starkware.starknet.testing.starknet import Starknet
 from utils import (
-    TestSigner, str_to_felt, TRUE, FALSE, get_contract_def, cached_contract, 
+    str_to_felt, TRUE, FALSE, get_contract_def, cached_contract, 
     assert_revert, to_uint
 )
+from signers import TestSigner
 
 
 signer = TestSigner(123456789987654321)
