+++ tests/token/erc721/test_ERC721_SafeMintable_mock.py
@@ -1,9 +1,10 @@
 import pytest
 from starkware.starknet.testing.starknet import Starknet
 from utils import (
-    TestSigner, str_to_felt, ZERO_ADDRESS, INVALID_UINT256, assert_revert,
+    str_to_felt, ZERO_ADDRESS, INVALID_UINT256, assert_revert,
     assert_event_emitted, get_contract_def, cached_contract, to_uint
 )
+from signers import TestSigner
 
 
 signer = TestSigner(123456789987654321)
