+++ tests/token/erc721_enumerable/test_ERC721_Enumerable_Mintable_Burnable.py
@@ -1,9 +1,10 @@
 import pytest
 from starkware.starknet.testing.starknet import Starknet
 from utils import (
-    TestSigner, str_to_felt, MAX_UINT256, get_contract_def, cached_contract,
+    str_to_felt, MAX_UINT256, get_contract_def, cached_contract,
     TRUE, assert_revert, to_uint, sub_uint, add_uint
 )
+from signers import TestSigner
 
 
 signer = TestSigner(123456789987654321)
